% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighCovBam.R
\name{HighCovBam}
\alias{\%HighCovBam\%}
\alias{HighCovBam}
\title{HighCovBam}
\usage{
HighCovBam(bam,cutoff=NULL,width=500)
}
\arguments{
\item{bam}{Bam file to analyze}
\item{cutoff}{cutoff threshold for minimum coverage. If NULL, it will be used 0.9 quantile}
\item{width}{size of windows. Default: 500 bases}
}
\value{
}
\description{
Function to find high coverage regions in a bam file.
}
\details{
}
\examples{
bam='example.bam'
HighCovBam(bam,cutoff=100,width=500)
}
