% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usePackage.R
\name{LiftoverVcf}
\alias{\%LiftoverVcf\%}
\alias{LiftoverVcf}
\title{LiftoverVcf}
\usage{
LiftoverVcf(vcf,output,from="hg38",to="hg19")
}
\arguments{
\item{vcf}{The input VCF/BCF file to be lifted overy}
\item{output}{name of the output file} 
\item{from}{starting assembly}
\item{to}{final assembly}
}
\value{
}
\description{
Function to lifts over a VCF file from one genome build to another, producing a properly headered, sorted VCF in one go.
}
\details{
}
\examples{
LiftoverVcf(vcf,output,from="hg38",to="hg19")
}
