% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SnpEff.R
\name{SnpEff}
\alias{\%SnpEff\%}
\alias{SnpEff}
\title{SnpEff}
\usage{
SnpEff(vcf,modifier=F,amino=F,rm_dup=T)
}
\arguments{
\item{vcf}{SnpEff-based annotated vcf file}
\item{modifier}{if TRUE variants with no impact on gene will be removed} 
\item{amino}{if TRUE variants that do not change the amino acid sequence will be removed}
\item{rm_dup}{if TRUE duplicated transcripts for each variant will be removed}
}
\value{
}
\description{
Function to make the SnpEff-based vcf readable by extracting the functional annotation and splitting each gene transcript into single rows.
}
\details{
}
\examples{
SnpEff(vcf,modifier=F,amino=F,rm_dup=T)
}
