% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitBed.R
\name{splitBed}
\alias{\%splitbed\%}
\alias{splitBed}
\title{splitBed}
\usage{
splitBed(bed,chunks=NULL,n=NULL,chrOnly=F,prefix=T,writeBed=F,verbose=F)
}
\arguments{
\item{bed}{data.frame in BED format with chromosome number, start and end position}
\item{chunks}{number of chunks to split}
\item{n}{number of lines for each chromosome}
\item{chrOnly}{split only by chromosome}
\item{prefix}{set to TRUE to add chr prefix if absent or FALSE to remove it if present}
\item{writeBed}{write new bed files}
\item{verbose}{give verbose output}
}
\value{
}
\description{
Function to split bed files by chunks, by number of rows for each chromosome or by chromosome
}
\details{
}
\examples{
splitBed(bed,n=100)
}
