% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usePackage.R
\name{usePackage}
\alias{\%uspckg\%}
\alias{usePackage}
\title{usePackage}
\usage{
usePackage(pkgs)
}
\arguments{
\item{pkgs }{Numeric vector (or single value) with the name of the packages}
}
\value{
}
\description{
Function to automatically check, install from both CRAN and Bioconductor repositories (if necessary) and load required packages
}
\details{
}
\examples{
usePackage(pkgs = c('data.table','Biobase'))
}
